services/language-service/pyproject.toml
[project]
name = "waqedi-language-service"
version = "0.1.0"
description = "Language Normalization & Multilingual Intelligence Service for WaqediAI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiokafka>=0.10.0",
    "langdetect>=1.0.9",
    "fasttext-wheel>=0.9.2",
    "regex>=2023.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
